+++ src/Helper/NativeCalls.cs
@@ -11,8 +11,6 @@ namespace SupercellProxy
         public const int MF_BYCOMMAND = 0x00000000;
         public const int SC_MAXIMIZE = 0xF030;
         public const int SC_SIZE = 0xF000;
-        //public const int SC_CLOSE = 0xF060;
-        //public const int SC_MINIMIZE = 0xF020;
     }
 
     class NativeCalls