+++ src/JSON Parser/JSONPacketWrapper.cs
@@ -2,30 +2,20 @@
 
 namespace SupercellProxy
 {
-    internal class JSONPacketWrapper
+    class JSONPacketWrapper
     {
         /// <summary>
-        ///     The ID of the Packet ex. 10100, 10101
+        /// The ID of the packet
         /// </summary>
         public int PacketID { get; set; }
 
         /// <summary>
-        ///     The Name of the Packet
+        /// The packetname
         /// </summary>
         public string PacketName { get; set; }
 
         /// <summary>
-        ///     Determines if this Packet should get Logged
-        /// </summary>
-        public bool ShouldLog { get; set; }
-
-        /// <summary>
-        ///     Determines if this Packet should get Exported
-        /// </summary>
-        public bool ShouldExport { get; set; }
-
-        /// <summary>
-        ///     The Fields of the Packet
+        /// Packet fields
         /// </summary>
         public List<JSONPacketField> Fields { get; set; }
     }