+++ src/Logger/LogType.cs
@@ -1,0 +1,15 @@
+﻿namespace SupercellProxy
+{
+    /// <summary>
+    /// Logging type
+    /// </summary>
+    enum LogType
+    {
+        INFO,
+        WARNING,
+        CONFIG,
+        PACKET,
+        EXCEPTION,
+        JSON
+    }
+}