+++ Changelog.md
@@ -1,10 +1,25 @@
 ﻿# Changelog
-# v1.0.1
+# v1.2.0
+	-JSON Parser:
+		-Code Cleanup
+		-Documentation Cleanup
+	-New packet export system (RawPackets, JsonPackets)
+	-Console command handling
+	-Custom exceptions
+	-Directory-Checks at startup
+	-New icon (Proxy.ico)
+	-Preview: Public Key Version Comparing
+	-Less unnecessary console output
+	-New console output format:
+		LOGTYPE - TEXT
+		
+# v1.1.0
     -Removed WebAPI
 	-JSON source implementation 
 	-New network code
 	-Logging improvements
 	-Code cleanup
+	
 # v1.0.0
     -All features from the old 1.5.0 version